feat: show points scored per answer
