+++ app/views/mashup_answers/_form.html.erb
@@ -20,7 +20,6 @@
   document.addEventListener("turbo:load", () => {
     const albumSelect = document.getElementById('album');
     const songSelect = document.getElementById('songs');
-    console.log(songSelect.dataset.content)
 
     updateSongOptions(albumSelect.value)
