+++ public/service_worker.js
@@ -1,13 +1,11 @@
 // public/service_worker.js
 
-self.addEventListener('install', (event) => {
-  console.log('Service Worker installing.');
+self.addEventListener('install', (_event) => {
   // Add a call to skipWaiting here if you want to immediately activate the service worker
   self.skipWaiting();
 });
 
-self.addEventListener('activate', (event) => {
-  console.log('Service Worker activating.');
+self.addEventListener('activate', (_event) => {
 });
 
 self.addEventListener('fetch', (event) => {
